backend/flask_proxy/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==2.2.5
Flask-Cors==4.0.1
idna==3.7
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.4
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
tqdm==4.66.4
typing-extensions==4.7.1
ubiops==4.4.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
